+++ CHANGELOG.md
@@ -1,5 +1,15 @@
+## 0.1.0-alpha.4
+- Přidáno získání a objednávání cizích jídel z burzy
+- Třída `Jidlo`: ~~cislo~~ 👉 **varianta**
+- Nová třída `Burza` pro cizí jídla z burzy
+- Více Exceptionů
+
+[Všechny změny](https://github.com/hernikplays/canteenlib/compare/0.1.0-alpha.3...0.1.0-alpha.4)
+
 ## 0.1.0-alpha.3
 - Kontrolovat správný status kód u GET požadavků
+
+[Všechny změny](https://github.com/hernikplays/canteenlib/compare/0.1.0-alpha.1...0.1.0-alpha.4)
 
 ## 0.1.0-alpha.2
 - Nevytvářet debugovací soubor